Add Elasticsearch settings and mapping types
Add a JSON file that can be PUT to Elasticsearch 5 when creating an index
to specify settings and mapping types.

This tries to follow what we had for Elasticsearch 0.90, with comments
in the JSON where it had to diverge.

This could be refactored as separate files for settings and mappings.
I'm not sure what's going to be most useful eventually. The current
format makes it easier to get started quickly with a new index by PUTing
the whole thing.

See
https://www.elastic.co/guide/en/elasticsearch/reference/5.3/indices-create-index.html