Add PropertyValueReport
This report takes a property and builds a Dataset that contains the values of the
property, the provider's local uri and the DPLA uri.

Refactors PropertyValueReport class name to PropertyDistinctValueReport because it is
only returning the unique values. Also to distinguish it from the PropertyValueReport
where all values for a given property are returned. This name change was suggested as
a part of DT-1498 (see Gretchen's comment)

Add test for selecting PropertyValueReport