Add all sourceResource properties to PropertyDistinctValueReport and PropertyValueReport.
Add extractValue to Report class that is used to extract the correct
value from the context case classes (SkosConcept, EdmAgent etc.). This
method also returns '__MISSING__' strings when the property is empty or
missing.

Adds case classes for each report (PropVal and PropDistinctVal) to support
extracting data and cleaning up the code.