Fix/id minting inversion (#90)
ID minting inversion

Addresses DT-1594. Moves the construction of the base id (what DPLA hashes)
into the Extractor trait and adds methods to be implemented in the provider's
extractor that provides the Extractor with all it needs to create the base id,
mint the DPLA id and construct the DPLA item URI.

Uses the shortName embedded in the harvest dataFrame for the provider prefix
(if that provider uses it to salt the local identifier). This method ensures
consistency and gate checks the provider short name before invoking the Extractor.

If the provider's identifier cannot be extracted an ExtractorException is raised.

Fixup MappingEntry for PA. Shortname in harvest DataFrame did not match name in mapping selector