Add tests, fix ParseDateEnrichment.preprocess
Add tests to document and verify behavior of this code, which sets the
stage for what the rest of the enrichment code will be receiving.

Change regular expressions in ParseDateEnrichment.process() for
assigning removedDecades and removedRanges. The old one for
removedDecades was not replacing "2010s" with "201x". The one for
removedRanges was changing "1978--" to "1978xx" but was not changing
"1901--".