Intial OkHttp refactor
Replace older request() methods scattered throughout the code with a single implementation
in HttpUtils. This touches a lot of other bits of code and this commit only addresses refactoring
in OAI, CDL, API, MDL and RS harvesters and the Twofishes calls.

This also removes the OaiQueryUrlBuilder and RsQueryUrlBuilder classes and moves the construction of OAI
URLs to the buildUrl() method in OaiResponseBuilder. Harvesters should now extend the UrlBuilder trait if
they need to to simplify the construction of request URLs

Removes a number of deprecated/unused utility methods and grungy code