stripBrackets enrichment [IN-350] (#144)
Adds string normalization member `stringBrackets`. This method removes a defined set of paired characters (e.g {}, [], ()) from a provided string. It constructs two regular expressions, one to search the string for a match and a second to identify characters to replace. Both of these patterns are built from a base pattern (see `baseBracketSearchPattern` and `baseBracketReplacePattern`) and the target characters passed in (see `brackets` member and `Bracket` case class). Both of the regex patterns ignore and replace extraneous white space respectively.